refactor(M2): extract upgrade scripts from index.js to src/upgrades.js
